src/backend/user-service/requirements.txt
# Base deps come from backend-base image.
# Add only user-service-specific packages here.
bcrypt==4.2.1
python-jose[cryptography]==3.5.0
sqlmodel==0.0.37
psycopg2-binary==2.9.10
